# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MediTrade API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/blogs": {
            "get": {
                "tags": [
                    "Blogs"
                ],
                "summary": "عرض كل المدونات",
                "operationId": "9819e5802278e060898795fe3c919599",
                "responses": {
                    "200": {
                        "description": "قائمة بكل المدونات"
                    }
                }
            },
            "post": {
                "tags": [
                    "Blogs"
                ],
                "summary": "إنشاء تدوينة جديدة",
                "operationId": "514d226035b3d74b6ccf9837b3d1ed36",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title_en",
                                    "title_ar",
                                    "short_description_en",
                                    "external_image",
                                    "internal_image",
                                    "header_image"
                                ],
                                "properties": {
                                    "title_en": {
                                        "type": "string"
                                    },
                                    "title_ar": {
                                        "type": "string"
                                    },
                                    "short_description_en": {
                                        "type": "string"
                                    },
                                    "short_description_ar": {
                                        "type": "string"
                                    },
                                    "full_description_en": {
                                        "type": "string"
                                    },
                                    "full_description_ar": {
                                        "type": "string"
                                    },
                                    "en_meta_title": {
                                        "type": "string"
                                    },
                                    "en_meta_description": {
                                        "type": "string"
                                    },
                                    "ar_meta_title": {
                                        "type": "string"
                                    },
                                    "ar_meta_description": {
                                        "type": "string"
                                    },
                                    "external_image": {
                                        "type": "file",
                                        "format": "binary"
                                    },
                                    "internal_image": {
                                        "type": "file",
                                        "format": "binary"
                                    },
                                    "header_image": {
                                        "type": "file",
                                        "format": "binary"
                                    },
                                    "external_image_alt_en": {
                                        "type": "string"
                                    },
                                    "external_image_alt_ar": {
                                        "type": "string"
                                    },
                                    "internal_image_alt_en": {
                                        "type": "string"
                                    },
                                    "internal_image_alt_ar": {
                                        "type": "string"
                                    },
                                    "header_image_alt_en": {
                                        "type": "string"
                                    },
                                    "header_image_alt_ar": {
                                        "type": "string"
                                    },
                                    "is_active": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "تم إنشاء التدوينة"
                    }
                }
            }
        },
...
        "/api/blogs/{id}": {
            "get": {
                "tags": [
                    "Blogs"
                ],
                "summary": "عرض تدوينة حسب ID",
                "operationId": "d5626a75caa63a57c4215c796839b85a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "رقم التدوينة",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "بيانات التدوينة"
                    },
                    "404": {
                        "description": "التدوينة غير موجودة"
                    }
                }
            },
            "post": {
                "tags": [
                    "Blogs"
                ],
                "summary": "تحديث تدوينة",
                "operationId": "2c3597f68790e0d30412c6cf92864869",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "رقم التدوينة",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "title_en": {
                                        "type": "string"
                                    },
                                    "title_ar": {
                                        "type": "string"
                                    },
                                    "short_description_en": {
                                        "type": "string"
                                    },
                                    "short_description_ar": {
                                        "type": "string"
                                    },
                                    "full_description_en": {
                                        "type": "string"
                                    },
                                    "full_description_ar": {
                                        "type": "string"
                                    },
                                    "en_meta_title": {
                                        "type": "string"
                                    },
                                    "en_meta_description": {
                                        "type": "string"
                                    },
                                    "ar_meta_title": {
                                        "type": "string"
                                    },
                                    "ar_meta_description": {
                                        "type": "string"
                                    },
                                    "external_image": {
                                        "type": "file",
                                        "format": "binary"
                                    },
                                    "internal_image": {
                                        "type": "file",
                                        "format": "binary"
                                    },
                                    "header_image": {
                                        "type": "file",
                                        "format": "binary"
                                    },
                                    "external_image_alt_en": {
                                        "type": "string"
                                    },
                                    "external_image_alt_ar": {
                                        "type": "string"
                                    },
                                    "internal_image_alt_en": {
                                        "type": "string"
                                    },
                                    "internal_image_alt_ar": {
                                        "type": "string"
                                    },
                                    "header_image_alt_en": {
                                        "type": "string"
                                    },
                                    "header_image_alt_ar": {
                                        "type": "string"
                                    },
                                    "is_active": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "تم التحديث بنجاح"
                    },
                    "404": {
                        "description": "التدوينة غير موجودة"
                    }
                }
            }
        },
        "/api/blogs/{id}/toggle-status": {
            "patch": {
                "tags": [
                    "Blogs"
                ],
                "summary": "تبديل حالة التدوينة (تفعيل / تعطيل)",
                "operationId": "3ece78723a882423dc3d37a9b23d9658",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "رقم التدوينة",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "تم تحديث الحالة بنجاح",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Status updated"
                                        },
                                        "is_active": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "التدوينة غير موجودة",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Blog not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "خطأ داخلي في الخادم",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Something went wrong"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Exception message here..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}